clean train entrance && preparing release new version
