+++ train_det.py
@@ -1,29 +1,20 @@
-#!/usr/bin/env python3
-# Copyright (c) Facebook, Inc. and its affiliates.
+"""
+train detection entrance
+
+Copyright @2022 YOLOv7 authors
 
 """
-TridentNet Training Script.
-
-This script is a simplified version of the training script in detectron2/tools.
-"""
-
 import os
-
 from detectron2.checkpoint import DetectionCheckpointer
 from detectron2.config import get_cfg
-from detectron2.engine import DefaultTrainer, default_argument_parser, default_setup, launch
+from detectron2.engine import DefaultTrainer, default_argument_parser, launch
 from detectron2.evaluation import COCOEvaluator
-from detectron2.data import MetadataCatalog, build_detection_train_loader, DatasetCatalog
-from detectron2.data.dataset_mapper import DatasetMapper
+from detectron2.data import MetadataCatalog, build_detection_train_loader
 from detectron2.modeling import build_model
 from detectron2.utils import comm
-import logging
-from detectron2.solver import build_lr_scheduler, LRMultiplier, WarmupParamScheduler
-from fvcore.common.param_scheduler import CosineParamScheduler
-
 from yolov7.data.dataset_mapper import MyDatasetMapper, MyDatasetMapper2
 from yolov7.config import add_yolo_config
-from yolov7.utils.allreduce_norm import all_reduce_norm
+from yolov7.utils.d2overrides import default_setup
 
 
 class Trainer(DefaultTrainer):
@@ -38,16 +29,12 @@ class Trainer(DefaultTrainer):
 
     @classmethod
     def build_train_loader(cls, cfg):
-        # return build_detection_train_loader(cfg, mapper=DatasetMapper(cfg, True))
-        # test our own dataset mapper to add more augmentations
         cls.custom_mapper = MyDatasetMapper2(cfg, True)
         return build_detection_train_loader(cfg, mapper=cls.custom_mapper)
 
     @classmethod
     def build_model(cls, cfg):
         model = build_model(cfg)
-        # logger = logging.getLogger(__name__)
-        # logger.info("Model:\n{}".format(model))
         return model
 
     def run_step(self):
@@ -58,25 +45,18 @@ class Trainer(DefaultTrainer):
         else:
             self.model.module.update_iter(self.iter)
 
-        if self.iter > self.cfg.INPUT.MOSAIC_AND_MIXUP.DISABLE_AT_ITER and self.cfg.INPUT.MOSAIC_AND_MIXUP.ENABLED:
+        if (
+            self.iter > self.cfg.INPUT.MOSAIC_AND_MIXUP.DISABLE_AT_ITER
+            and self.cfg.INPUT.MOSAIC_AND_MIXUP.ENABLED
+        ):
             # disable augmentation
             self.cfg.defrost()
             self.cfg.INPUT.MOSAIC_AND_MIXUP.ENABLED = False
             self.cfg.freeze()
             self.custom_mapper.disable_aug()
 
-        # if comm.is_main_process():
-        #     # when eval period, apply all_reduce_norm as in https://github.com/Megvii-BaseDetection/YOLOX/issues/547#issuecomment-903220346
-        #     interval = self.cfg.SOLVER.CHECKPOINT_PERIOD if self.cfg.TEST.EVAL_PERIOD == 0 else self.cfg.TEST.EVAL_PERIOD
-        #     if self.iter % interval == 0 and self.iter != 0:
-        #         all_reduce_norm(self.model)
-        #         self.checkpointer.save('latest')
-
 
 def setup(args):
-    """
-    Create configs and perform basic setups.
-    """
     cfg = get_cfg()
     add_yolo_config(cfg)
     cfg.merge_from_file(args.config_file)
@@ -88,7 +68,6 @@ def setup(args):
 
 def main(args):
     cfg = setup(args)
-
     if args.eval_only:
         model = Trainer.build_model(cfg)
         DetectionCheckpointer(model, save_dir=cfg.OUTPUT_DIR).resume_or_load(
@@ -99,15 +78,11 @@ def main(args):
 
     trainer = Trainer(cfg)
     trainer.resume_or_load(resume=args.resume)
-    # print('trainer.start: ', trainer.start_iter)
-    # trainer.model.iter = trainer.start_iter
-    # print('trainer.start: ', trainer.model.iter)
     return trainer.train()
 
 
 if __name__ == "__main__":
     args = default_argument_parser().parse_args()
-    print("Command Line Args:", args)
     launch(
         main,
         args.num_gpus,
