+++ train_inseg.py
@@ -1,28 +1,29 @@
-#!/usr/bin/env python3
-# Copyright (c) Facebook, Inc. and its affiliates.
-"""
-
-Training script using custom coco format dataset
-
-what you need to do is simply change the img_dir and annotation path here
-Also define your own categories.
-
-"""
-
 import os
 from detectron2.checkpoint import DetectionCheckpointer
 from detectron2.config import get_cfg
-from detectron2.engine import DefaultTrainer, default_argument_parser, default_setup, launch
+from detectron2.engine import (
+    DefaultTrainer,
+    default_argument_parser,
+    default_setup,
+    launch,
+)
 from detectron2.evaluation import COCOEvaluator
-from detectron2.data import MetadataCatalog, build_detection_train_loader, DatasetCatalog
-from detectron2.data.datasets.coco import load_coco_json, register_coco_instances
-from detectron2.data.dataset_mapper import DatasetMapper
+from detectron2.data import (
+    MetadataCatalog,
+    build_detection_train_loader,
+    DatasetCatalog,
+)
 from detectron2.modeling import build_model
 from detectron2.utils import comm
 
 from yolov7.config import add_yolo_config
 from yolov7.data.dataset_mapper import MyDatasetMapper2, MyDatasetMapper
 from yolov7.utils.allreduce_norm import all_reduce_norm
+
+
+"""
+This using for train instance segmentation!
+"""
 
 
 class Trainer(DefaultTrainer):
@@ -34,14 +35,14 @@ class Trainer(DefaultTrainer):
 
     @classmethod
     def build_train_loader(cls, cfg):
-        return build_detection_train_loader(cfg,
-                                            mapper=MyDatasetMapper(cfg, True))
+        # return build_detection_train_loader(cfg,
+        #                                     mapper=MyDatasetMapper(cfg, True))
+        cls.custom_mapper = MyDatasetMapper(cfg, True)
+        return build_detection_train_loader(cfg, mapper=cls.custom_mapper)
 
     @classmethod
     def build_model(cls, cfg):
         model = build_model(cfg)
-        # logger = logging.getLogger(__name__)
-        # logger.info("Model:\n{}".format(model))
         return model
 
     def run_step(self):
@@ -54,9 +55,6 @@ class Trainer(DefaultTrainer):
 
 
 def setup(args):
-    """
-    Create configs and perform basic setups.
-    """
     cfg = get_cfg()
     add_yolo_config(cfg)
     cfg.merge_from_file(args.config_file)
@@ -72,26 +70,23 @@ def main(args):
     if args.eval_only:
         model = Trainer.build_model(cfg)
         DetectionCheckpointer(model, save_dir=cfg.OUTPUT_DIR).resume_or_load(
-            cfg.MODEL.WEIGHTS, resume=args.resume)
+            cfg.MODEL.WEIGHTS, resume=args.resume
+        )
         res = Trainer.test(cfg, model)
         return res
 
     trainer = Trainer(cfg)
     trainer.resume_or_load(resume=args.resume)
-    # print('trainer.start: ', trainer.start_iter)
-    # trainer.model.iter = trainer.start_iter
-    # print('trainer.start: ', trainer.model.iter)
     return trainer.train()
 
 
 if __name__ == "__main__":
     args = default_argument_parser().parse_args()
-    print("Command Line Args:", args)
     launch(
         main,
         args.num_gpus,
         num_machines=args.num_machines,
         machine_rank=args.machine_rank,
         dist_url=args.dist_url,
-        args=(args, ),
+        args=(args,),
     )
